Rename cmd options source files.
