+++ src/app/CMakeLists.txt
@@ -3,14 +3,14 @@ include_directories(${CMAKE_CURRENT_BINARY_DIR})
 
 set(QBT_APP_HEADERS
 application.h
+cmdoptions.h
 filelogger.h
-options.h
 )
 
 set(QBT_APP_SOURCES
 application.cpp
+cmdoptions.cpp
 filelogger.cpp
-options.cpp
 main.cpp
 )
